\name{head-methods}
\docType{methods}
\alias{head-methods}
\alias{head,ANY-method}
\alias{head,DirectedHyperedge-method}
\title{Methods for Function head in Package hypergraph}
\description{
 Returns the nodes that are in the head of the hyperedge.
}
\section{Methods}{
\describe{

\item{x = "DirectedHyperedge"}{ Return the nodes that are in the head of
  the directed hyperedge}
}}
\seealso{
  \code{\link{tail-methods}}
  \code{\link{DirectedHyperedge-class}}
}
\keyword{methods}

