\name{inciMat-methods}
\docType{methods}
\alias{inciMat-methods}
\alias{inciMat,Hypergraph-method}
\title{  Methods for Function inciMat in Package hypergraph }
\description{
  Methods for function \code{inciMat} in package \pkg{hypergraph} 
}
\section{Methods}{
\describe{

\item{.Object = "Hypergraph"}{Return the incidence matrix representation
of the hypergraph.  The rows the matrix represent the nodes and the
columns represent the hyperedges.}
}}
\keyword{methods}

