\name{label-methods}
\docType{methods}
\alias{label-methods}
\alias{label,gNode-method}
\alias{label,Hyperedge-method}
\title{Methods for Function label in Package hypergraph}
\description{
 The \code{label} method provides access to the text label of
 \code{\link{Hyperedge-class}} objects.
}
\section{Methods}{
\describe{
\item{object = "Hyperedge"}{ Return the character string attribute of
the object.}
}}
\keyword{methods}

