\name{nodes-methods}
\docType{methods}
\alias{nodes-methods}
\alias{nodes,graphNEL-method}
\alias{nodes,distGraph-method}
\alias{nodes,clusterGraph-method}
\alias{nodes,graphH-method}
\alias{nodes,Hyperedge-method}
\alias{nodes,DirectedHyperedge-method}
\alias{nodes,Hypergraph-method}
\title{Methods for Function nodes in Package hypergraph}
\description{
 Methods for function \code{nodes} in package \pkg{hypergraph}
}
\section{Methods}{
\describe{
\item{object = "Hypergraph"}{Return a vector containing the nodes in the
  hypergraph}
\item{object = "Hyperedge"}{Return a vector containing the nodes in the
  hyperedge}
\item{object = "DirectedHyperedge"}{Return a vector containing the nodes in the
  directed hyperedge}
}}
\keyword{methods}

