\name{numNodes-methods}
\docType{methods}
\alias{numNodes-methods}
\alias{numNodes,graph-method}
\alias{numNodes,graphNEL-method}
\alias{numNodes,distGraph-method}
\alias{numNodes,clusterGraph-method}
\alias{numNodes,Hypergraph-method}
\title{Methods for Function numNodes in Package hypergraph}
\description{
 Methods for function \code{numNodes} in package \pkg{hypergraph}
}
\section{Methods}{
\describe{

\item{object = "Hypergraph"}{Return the number of nodes in the hypergraph}
}}
\keyword{methods}

