\name{tail-methods}
\docType{methods}
\alias{tail-methods}
\alias{tail,ANY-method}
\alias{tail,DirectedHyperedge-method}
\title{Methods for Function tail in Package hypergraph}
\description{
 Methods for function \code{tail} in package \pkg{hypergraph}
}
\section{Methods}{
\describe{
\item{x = "DirectedHyperedge"}{Return a vector of nodes in the tail of
the directed hyperedge}
}}
\keyword{methods}

