\name{toGraphNEL-methods}
\docType{methods}
\alias{toGraphNEL-methods}
\alias{toGraphNEL,Hypergraph-method}
\title{  Methods for Function toGraphNEL in Package hypergraph }
\description{
  Methods for function \code{toGraphNEL} in package \pkg{hypergraph} 
}
\section{Methods}{
\describe{

\item{.Object = "Hypergraph"}{Convert the Hypergraph to a bipartite
  graph represented by a \code{\link{graphNEL}} object.}
}}
\keyword{methods}

